speech-ai-backend/requirements.txt
Flask
openai
openai-whisper
pydub
